requirements.txt
dj-database-url==0.5.0
gunicorn==20.1.0
psycopg2==2.8.6
whitenoise==5.2.0
social-auth-app-django
pyOpenSSL==20.0.1
werkzeug==2.0.1
django-extensions==3.1.3
itsdangerous